Add a simulator device.
Use this to get serial working in the simulator.
You now 'flash' the simulator to run code.
