+++ src/App.tsx
@@ -10,7 +10,7 @@ import "./App.css";
 import { DialogProvider } from "./common/use-dialogs";
 import VisualViewPortCSSVariables from "./common/VisualViewportCSSVariables";
 import { deployment, useDeployment } from "./deployment";
-import { MicrobitWebUSBConnection } from "./device/device";
+import { MicrobitWebUSBConnection } from "./device/webusb";
 import {
   DeviceContextProvider,
   SyncStatusProvider,
